backend/requirements.txt
flask
flask-cors
bcrypt
gunicorn
mysql-connector-python